+++ packages/components/scripts/generateDSRComponents.ts
@@ -424,7 +424,8 @@ import { get${componentName}Css } from '${stylesBundleImportPath}';
       } else if (tagName === 'p-scroller') {
         newFileContent = newFileContent
           .replace(/(this\.)props\.(is(?:Next|Prev)Hidden)/g, '$1$2')
-          .replace(/(deprecationMap\[this\.props\.gradientColorScheme)/, '$1 as ScrollerGradientColorScheme');
+          .replace(/(deprecationMap\[this\.props\.gradientColorScheme)/, '$1 as ScrollerGradientColorScheme')
+          .replace(/(deprecationMap\[this\.props\.gradientColor)/, '$1 as ScrollerGradientColor');
       } else if (tagName === 'p-popover') {
         // only keep :host , button, .icon & .label styles
         newFileContent = newFileContent