+++ packages/components/CHANGELOG.md
@@ -31,6 +31,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0), 
   ([#3953](https://github.com/porsche-design-system/porsche-design-system/pull/3953))
 - `Multi Select`: aligned filter and keyboard behavior with `Select`
   ([#3943](https://github.com/porsche-design-system/porsche-design-system/pull/3943))
+- `Scroller`: Deprecation of prop `gradientColor` because gradient is now managed internally and adapts to all themes
+  and background colors automatically.
+  ([#3962](https://github.com/porsche-design-system/porsche-design-system/pull/3962))
 
 ## [3.29.0-rc.3] - 2025-07-29